ML: Patch for MLPrediction/MinConfs/NewConfs (prevent few samples)

+++ aloja-web/Controller/MLPredictionController.php
@@ -198,6 +198,9 @@ class MLPredictionController extends AbstractController
 				fputcsv($fp,array_values(array_merge($exec_names,$net_names,$disk_names,$bench_names)),',','"');
 			    	foreach($rows as $row) fputcsv($fp, array_values($row),',','"');
 
+				// Check we have enough values
+				if (count($rows) < 10) throw new \Exception('WARNING: Too many few samples selected to learn ('.count($rows).'). Change your filter to use a wider data slice.');
+
 				// run the R processor
 				exec('cd '.getcwd().'/cache/ml ; touch '.getcwd().'/cache/ml/'.md5($config).'.lock');
 				exec('cd '.getcwd().'/cache/ml ; '.getcwd().'/resources/queue -c "'.getcwd().'/resources/aloja_cli.r -d '.$cache_ds.' -m '.$learn_method.' -p '.$learn_options.' > /dev/null 2>&1; rm -f '.getcwd().'/cache/ml/'.md5($config).'.lock; touch '.md5($config).'.fin" > /dev/null 2>&1 -p 1 &');
